LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/device-pixel-adjustment-expected.txt


FAIL Test that scroll anchor adjustments can happen by a sub device-pixel amount. assert_equals: adjusting by 0.5px, 2 times, should be the same as adjusting by 1px, once. expected 11 but got 10
